마침내 등록에 성공하였도다

+++ src/main/java/com/os/dto/InsertDTO.java
@@ -18,7 +18,7 @@ public class InsertDTO {
     private String paymentTitle;
     private OrderType paymentType;
     private BizTo paymentBizTo;
-    private Long paymentMonth;
+    private int paymentMonth;
     private int autoDate;
     private int paymentFirstPay;
     private List<ProductDTO> productList;
